+++ ChangeLog.md
@@ -5,3 +5,4 @@ v 0.1.1.0
 * Documentation Fixes
 * Added export of newPandoc and NamedDoc to Knit.Report to facilitate multi-doc use.
 * Removed redundant imports from Knit.Report.Input.Table.Colonnade
+* Added a multi-doc example